Add Health & Safety page
This is based on the verbiage sent around by Elna in a Google Doc shared with
the team.

This also makes the CSS around the header slightly less horrible to accommodate
more items better, and it changes the header styles a bit to make the H&S page
look ok.

Signed-off-by: Julius Volz <julius.volz@gmail.com>